Summary: format -- linking primitives Keywords:
Documentation for label, reference, pageref and include.

Add a paragraph to the "fundamental primitives" page to define "line
context", "line content", and "block content" which are important to
understand where "include" is allowed.

Edit the documentation for "cell" to define a "multiparagraph cell"
and emphasize the exceptional nature of this construct.

Patches applied:

 * ddaa@ddaa.net--2004-texmacs/doc--devo--1.0.4--patch-27
   format -- linking primitives


git-svn-id: svn+ssh://svn.sv.gnu.org/texmacs/trunk@336 64cb5145-927a-446d-8aed-2fb7b4773692